[WIP] mbpt::Operator::commutes_with was not correct
in fact design of ParticleNumberChange needs to change to encode number of creators/annihilators in each relevant space. Then 2 ops do not commute if one has one or more annihilators in a space while the other has one or more creators in that same space. This commit implements this logic seemingly correctly but breaks canonicalize() somehow by causing infinite loop in bubble sort