Prevent optimize::reorder from swallowing terms
If the expression fed into the optimize function is a sum that contains
an addend which is either a sum or a product itself, but only contains a
single element (e.g. a single addend or a single factor), the visiting
mechanism inside the cluster function would end up not visiting those
terms and thereby preventing them from being processed (aka: they are
swallowed and never see the light of day again).

A test case checking for the expected (non-swallowing) behavior is added
as well.