Get rid of duplication in testing of Sum equality
Instead of hardcoding two expected LaTeX representations of a given
expression, we now do an exhaustive search over all permutation of
summands to see if one of them yields the desired permutation.
This is acceptable as the order of summands has no practical relevance
other than for matching a hard-coded order in a test case.

The benefit of doing it this way is reduced maintenance cost in exchange
for a runtime performance overhead when running the tests. However, the
overhead is expected to be small and for test cases it shouldn't matter
anyway.