Revert "TensorBlockIndexComparer uses Index::operator< which is the right way to compare Indices"
This reverts commit ba832da3054fc181fee88d7c45f627a5ea9f1ae7.

The reason being that Index::operator< is NOT the right way to compare
indices in this context. If we also account for index labels (instead of
only their space), we are comparing full indexed tensor elements rather
than tensor blocks (slots) as this class is supposed to do.

In order for test cases to work with this re-introduced change, this
commit also reverts
"amend tests for ba832da3054fc181fee88d7c45f627a5ea9f1ae7"
(commit 4a9543851105e2bae0939495986351fe74930d70)

Finally, a hand full of additional test cases had to be adapted to the
changed canonicalization routine.