Get rid of grand_index_list_ member variable
The type of this member variable being the type that it was caused
compiler warnings. Also, it turns out that there is no need for this
index collection to be a member variable in the first place. It is only
ever used very locally in the code. Hence, it has now been converted
into a local variable instead. As a bonus, this gets rid of the compiler
warning as the member variable it referred to no longer exists.

Note that not caching the index list should be fine as the functions it
is being used in are almost never called repeatedly on the same
TensorNetwork instance. Therefore, the caching benefit that the member
variable provided is nullified.

Fixes #250