Refactor deprecated use of shared_ptr atomic access API
Using a plain std::shared_ptr together with std::atomic_{load,store} has
been deprecated in C++20 and the API is removed in C++26. See P2869
(and/or https://github.com/cplusplus/papers/issues/1525) for details.

This commit removes the use of this depracted API in the way suggested
by P2869 (that is, wrap in a std::atomic type and use that directly
rather than via std::atomic_{load,store}).