Avoid self-consistent index replacement for permutations
Trying to apply a given index replacement in a loop until the expression
no longer changes will cause an endless loop if any of the replacements
replaces an index in the target expression with one that contains a
proto index of the same name. So e.g. i1 -> a1<i1>
In this case, in every iteration, the i1 proto index of a1<i1> will
again be replaced by a1<i1>, leading to nested proto indices as well as
the aforementioned endless loop due to infinitely deep nested proto
indices.