Fix spintrace skipping contributing terms
In cases where spintrace was given a Product, but tracing that product
would yield a single Tensor, this contribution would be ignored.

To fix this, the check whether or not to include the traced term is now
explicitly checking for whether the term is zero (instead of checking
the term's size to be zero, which is true for a Tensor). If it is not,
the term is included.