Remove Generator::insert_blank_lines
This function was odd as it was the only function that didn't encode any
semantic operation. Also, its use lead to spurious blank lines for
generators that don't need to declare all indices/variables/tensors.

The current use of this function has been replaced with semantic
callbacks informing the generator that the declarations of a given type
are finished now. If it wants to, the generator can initiate an
appropriate action on being notified about this.