Implement tree traversing as loop
Using recursive function calls makes for easier code but it also sets an
implicit limit on the maximum tree depth before the program will crash
due to a stack overflow. The loop-based version doesn't have this
problem. This limit is encountered for real applications (e.g. icMRCCSD
which has thousands of contributions for a single result).

Additionally, the loop-based implementation can be expected to be (much)
more performant in cases in which the compiler is not able to rewrite
the recursive calls into a loop itself. Though, the performance
difference is likely of no practical importance.