Don't let GenerationOptimizer inherit from main generator
This only leads to having to fight against virtual dispatch which we
don't want when the respective function is called from the optimizer.
This becomes hopeless as soon as the base generator delegates function
calls to other function calls of itself.

Hence, we know explicitly wrap a generator object of the desired type
and forward all calls to that.