cmake: `doc/examples` are now executed and tested
If testing is enabled, the selected examples are added to the test suite. If the executable produces a runtime error, corresponding test will fail.
Also introduced an option to disable time-consuming examples. If the source file contains // SEQUANT_DOC_EXAMPLE: DO NOT RUN`, the file will not be added to the test suite.