CI: Don't install Catch2 via homebrew
Since Catch v3.8, the discovery of the tests happens via the built-in
JSON reporter. The (apparent) issue with that is that any output that
gets printed to stdout that is not captured by Catch2 (e.g. anything
that is printed before Catch takes over in main()) will mess with the
printed JSON, leading to invalid syntax and subsequent failure of test
discovery.

For tests in SeQuant, there seems to be such problematic output and the
version installed via brew is >= v3.8 so that this is actually an issue.

See also:
- https://github.com/catchorg/Catch2/releases/tag/v3.8.0
- https://discourse.cmake.org/t/using-catch2-test-detection-when-test-binary-is-output-outside-the-build-tree/13340/6