Fix wrong vertex determination in TNv2::canonicalize_slots
The old code was assuming that vertices in the graph follow the same
scheme as in TNv1 (indices being the first vertices). However, this is
not applicable to TNv2. Instead, we have to make use of the provided
Index-to-Vertex map (which also eliminates any implicit assumptions
about the vertex order).