Undo indices() -> braketaux() renaming
This renaming has changed the semantics of the code in a way that would
cause bugs in the future, in case there will be another index group in
addition to bra, ket and aux.

Additionally, AbstractTensor has gained a proper interface for indices
and in the spirit of functions like braketaux_indices we now distinguish
between indices() and slots().