Better error handling
Instead of only calling abort, we use an assert that will always fail
(if enabled) which contains an error message (and will report the exact
spot where failure happened). In cases in which assertions are disabled,
we print the error message to the console instead.

Additionally, unreachable code has now been marked by a dedicated macro,
which will use compiler builtins (or C++23 features) to tell the
compiler about this unreachability. Only in cases where we don't have a
standard or compiler-specific way of encoding this, do we fall back to
doing a plain abort.
If assertions are enabled, we'll trigger an assertion if code takes the
supposedly unreachable path.

Fixes #122