Get rid of sequant/core/any.hpp
The wrapping logic in that file was broken in such a way that std::any
was never chosen. In fact, fixing the logic leads to compile failures
because the code in tag.hpp is written in a way to only define the
Taggable class (required by e.g. Index) if std::any is supposedly not
found.

Hence, this commit removes the any header and refactors the tag header
to be able to live without it.

As a nice side-effect, this fixes a compiler warning from GCC 15 due to
the use of the ciso646 header file, which (formally) got removed with
C++20.