Benchmark Comparison Workflow (#331)
* benchmark: add an alternate benchmark target to just run cc benchmarks

* benchmark: reduce max CC rank to 15

* locale: can optionally disable thousands operator

Why: When writing JSON/CSV, the standard locale set by `set_locale` breaks the JSON/CSV structure because of the presence of commas in large numbers.

* benchmark: disable thousands separator in benchmark output

Presence of commas in numbers break JSON/CSV output

* benchmark: introduce python script for benchmark comparison

This script can be used to compare peformance of two commits. It checks out, builds and compares benchmark outputs of both commits.

How to use:
python3 benchmark_compare.py abc123 def456
python3 benchmark_compare.py master your/feature/branch

* ci: introduce benchmark comparison workflow

This is a workflow which can be triggered by commenting "\benchmark" on PRs (or through manual dispatch).

It uses the `benchmark_compare.py` to run and compare benchmarks. The comparison result is posted as a comment on the PR.

* ci: avoid duplicate benchmark workflows

* ci: remove some debug info [skip ci]

* benchmark: improve filename handling, make `compare_benchmarks` reusable

- When trying to benchmark with branch names, presence of '\' can have issues, replace them with '-'
- output filename changed to "benchmark_comparison.txt"
- Add ability to import the compare_benchmarks function for independent use

* ci: update output file name in benchmark_compare.yml

* benchmark: enforce release build, even though it is the default

* fix typo

* benchmark: several fixes, ccache, faster builds, slightly better error handling

* locale: drop the thousands operator logic, not needed since we switched default locale

* benchmark: drop unused cc_benchmark target + tabs fix [skip ci]

* fix: revert changes to runtime files

* ci: simplify CMake options in benchmark script [skip ci]

Eval tests are off if tests are off. Benchmarks run on single-threads, no need for Context thread safety option.

* ci: drop unused logic in benchmark script [skip ci]

---------

Co-authored-by: Eduard Valeyev <eduard@valeyev.net>